src/test/java/com/mageddo/dnsproxyserver/dnsconfigurator/LinuxDnsConfiguratorTest.java
package com.mageddo.dnsproxyserver.dnsconfigurator;

import com.mageddo.dnsproxyserver.server.dns.IP;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.io.TempDir;

import java.nio.file.Files;
import java.nio.file.Path;

import static org.junit.jupiter.api.Assertions.assertEquals;

class LinuxDnsConfiguratorTest {

  LinuxDnsConfigurator configurator = new LinuxDnsConfigurator();

  @Test
  void mustConfigureDpsServerOnEmptyFile(@TempDir Path tmpDir) throws Exception {

    // arrrange
    final var resolvFile = Files.createTempFile(tmpDir, "resolv", ".conf");

    // act
    this.configurator.configure(IP.of("10.10.0.1"), resolvFile);

    // assert
    assertEquals(
      """
        nameserver 10.10.0.1 # dps-entry
        """,
      Files.readString(resolvFile)
    );

  }
}
